+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -59,6 +59,7 @@ This release contains a number of fairly significant changes. The focus has been
     * `InvertVector3Processor`
     * `NormalizeVector2Processor`
     * `NormalizeVector3Processor`
+- Added `MultiTapInteraction`. Can be used to listen for double-taps and the like.
 
 ### Fixes